Forgot to use a cast in cdf().

+++ b4o/src/sonumina/b4o/calculation/EmpiricalDistribution.java
@@ -36,7 +36,7 @@ public class EmpiricalDistribution
 	public double cdf(double x, boolean lowerTail)
 	{
 		int idx = Arrays.binarySearch(observations, x);
-		return idx/observations.length;
+		return idx/(double)observations.length;
 	}
 	
 }
